+++ examples/client/config/agnss.cpp
@@ -1,0 +1,148 @@
+#include <loglet/loglet.hpp>
+#include "../config.hpp"
+
+#include <inttypes.h>
+
+#define LOGLET_CURRENT_MODULE &LOGLET_MODULE_REF2(client, config)
+
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wsuggest-destructor-override"
+#pragma GCC diagnostic ignored "-Wdeprecated-copy-with-user-provided-dtor"
+#pragma GCC diagnostic ignored "-Wnewline-eof"
+#pragma GCC diagnostic ignored "-Wmissing-variable-declarations"
+#pragma GCC diagnostic ignored "-Winconsistent-missing-destructor-override"
+#pragma GCC diagnostic ignored "-Wsuggest-override"
+#pragma GCC diagnostic ignored "-Wshadow-field"
+#pragma GCC diagnostic ignored "-Wsuggest-destructor-override"
+#include <args.hpp>
+#pragma GCC diagnostic pop
+
+namespace agnss {
+
+static args::Group                  gGroup{"A-GNSS:"};
+static args::Flag gEnable{
+    gGroup,
+    "enable",
+    "Enable A-GNSS client",
+    {"agnss-enable"},
+};
+static args::ValueFlag<std::string> gHost{
+    gGroup,
+    "host",
+    "A-GNSS server hostname or IP address",
+    {"agnss-host"},
+    args::Options::Single,
+};
+static args::ValueFlag<uint16_t> gPort{
+    gGroup, "port", "A-GNSS server port", {"agnss-port"}, args::Options::Single,
+};
+static args::ValueFlag<std::string> gInterface{
+    gGroup, "interface", "Interface to use for A-GNSS server connection",
+    {"agnss-interface"}, args::Options::Single,
+};
+static args::Flag gGps{
+    gGroup,
+    "gps",
+    "Request GPS assistance data",
+    {"agnss-gps"},
+};
+static args::Flag gGlonass{
+    gGroup,
+    "glonass",
+    "Request GLONASS assistance data",
+    {"agnss-glonass"},
+};
+static args::Flag gGalileo{
+    gGroup,
+    "galileo",
+    "Request Galileo assistance data",
+    {"agnss-galileo"},
+};
+static args::Flag gBeidou{
+    gGroup,
+    "beidou",
+    "Request BeiDou assistance data",
+    {"agnss-beidou"},
+};
+static args::ValueFlag<long> gInterval{
+    gGroup, "interval", "Request interval in seconds", {"agnss-interval"}, 60,
+};
+static args::ValueFlag<uint64_t> gMsisdn{
+    gGroup, "msisdn", "MSISDN identity", {"agnss-msisdn"}, args::Options::Single,
+};
+static args::ValueFlag<uint64_t> gImsi{
+    gGroup, "imsi", "IMSI identity", {"agnss-imsi"}, args::Options::Single,
+};
+static args::ValueFlag<std::string> gIpv4{
+    gGroup, "ipv4", "IPv4 address identity", {"agnss-ipv4"}, args::Options::Single,
+};
+
+void setup(args::ArgumentParser& parser) {
+    static args::GlobalOptions globals{parser, gGroup};
+}
+
+void parse(Config* config) {
+    auto& agnss = config->agnss;
+    agnss.enabled = gEnable.Get();
+    
+    if (!agnss.enabled) {
+        return;
+    }
+
+    if (!gHost) {
+        throw args::RequiredError("`--agnss-host` is required when A-GNSS is enabled");
+    }
+    if (!gPort) {
+        throw args::RequiredError("`--agnss-port` is required when A-GNSS is enabled");
+    }
+
+    agnss.host = gHost.Get();
+    agnss.port = gPort.Get();
+
+    if (gInterface) {
+        agnss.interface = std::unique_ptr<std::string>(new std::string(gInterface.Get()));
+    }
+
+    agnss.gps = gGps.Get();
+    agnss.glonass = gGlonass.Get();
+    agnss.galileo = gGalileo.Get();
+    agnss.beidou = gBeidou.Get();
+    agnss.interval_seconds = gInterval.Get();
+
+    if (gMsisdn) {
+        agnss.msisdn = std::unique_ptr<uint64_t>(new uint64_t(gMsisdn.Get()));
+    }
+    if (gImsi) {
+        agnss.imsi = std::unique_ptr<uint64_t>(new uint64_t(gImsi.Get()));
+    }
+    if (gIpv4) {
+        agnss.ipv4 = std::unique_ptr<std::string>(new std::string(gIpv4.Get()));
+    }
+}
+
+void dump(AGnssConfig const& config) {
+    DEBUGF("enabled: %s", config.enabled ? "true" : "false");
+    if (!config.enabled) return;
+
+    DEBUGF("host: \"%s\"", config.host.c_str());
+    DEBUGF("port: %d", config.port);
+    if (config.interface) {
+        DEBUGF("interface: \"%s\"", config.interface.get()->c_str());
+    }
+    DEBUGF("gps: %s", config.gps ? "true" : "false");
+    DEBUGF("glonass: %s", config.glonass ? "true" : "false");
+    DEBUGF("galileo: %s", config.galileo ? "true" : "false");
+    DEBUGF("beidou: %s", config.beidou ? "true" : "false");
+    DEBUGF("interval: %ld seconds", config.interval_seconds);
+    if (config.msisdn) {
+        DEBUGF("msisdn: %" PRIu64, *config.msisdn);
+    }
+    if (config.imsi) {
+        DEBUGF("imsi: %" PRIu64, *config.imsi);
+    }
+    if (config.ipv4) {
+        DEBUGF("ipv4: \"%s\"", config.ipv4->c_str());
+    }
+}
+
+}  // namespace agnss